Change PrintInt to Take Long Long

+++ main.h
@@ -5,9 +5,9 @@
 #include <stdlib.h>
 
 int printChar(char);
-	int _printf(const char *format, ...);
-	int printDouble(float);
-	int printString(char *);
-	int printInt(int);
+int _printf(const char *format, ...);
+int printDouble(float);
+int printString(char *);
+int printInt(long long);
 
 #endif
